requirements.txt
selenium
beautifulsoup4
requests
python-telegram-bot
pyyaml
black